Added KedroDataCatlog tests template
